requirements.txt
python-telegram-bot==20.3
spacy==3.7.2
python-dotenv==1.0.1

# Pin a compatible small English model. If this URL 404s for your platform, install via `python -m spacy download en_core_web_sm`.
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

